Add code.rs and share the application logic of the markdown example with code.rs
