+++ shell/examples/code.rs
@@ -1,0 +1,4 @@
+#[path = "shared/application.rs"]
+mod application;
+
+fn main() {}
