+++ shell/examples/shared/application.rs
@@ -1,0 +1,225 @@
+use std::collections::HashMap;
+
+use massive_geometry::{Camera, Matrix4, Point, PointI, SizeI, Vector3};
+use massive_shapes::{GlyphRun, Shape};
+use shell::Shell;
+use winit::{
+    event::{
+        DeviceId, KeyEvent, Modifiers, MouseButton, MouseScrollDelta, TouchPhase, WindowEvent,
+    },
+    keyboard::{Key, NamedKey},
+};
+
+use massive_shell as shell;
+
+pub struct Application {
+    camera: Camera,
+    glyph_runs: Vec<(Point, GlyphRun)>,
+    page_size: SizeI,
+
+    /// If pressed, the origin.
+    left_mouse_button_pressed: Option<LeftMouseButtonPressed>,
+    middle_mouse_button_pressed: Option<MiddleMouseButtonPressed>,
+    /// Tracked positions of all devices.
+    positions: HashMap<DeviceId, PointI>,
+    modifiers: Modifiers,
+
+    /// Current x / y Translation.
+    translation: PointI,
+    translation_z: i32,
+    /// Rotation in discrete degrees.
+    rotation: PointI,
+}
+
+impl Application {
+    pub fn new(
+        camera: Camera,
+        glyph_runs: Vec<(Point, GlyphRun)>,
+        page_size: impl Into<SizeI>,
+    ) -> Self {
+        Self {
+            camera,
+            glyph_runs,
+            page_size: page_size.into(),
+
+            left_mouse_button_pressed: None,
+            middle_mouse_button_pressed: None,
+            positions: HashMap::new(),
+            modifiers: Modifiers::default(),
+            translation: PointI::default(),
+            translation_z: 0,
+            rotation: PointI::default(),
+        }
+    }
+}
+
+struct LeftMouseButtonPressed {
+    origin: PointI,
+    translation_origin: PointI,
+}
+
+struct MiddleMouseButtonPressed {
+    origin: PointI,
+    rotation_origin: PointI,
+}
+
+const MOUSE_WHEEL_PIXEL_DELTA_TO_Z_PIXELS: f64 = 0.25;
+const MOUSE_WHEEL_LINE_DELTA_TO_Z_PIXELS: i32 = 16;
+
+impl shell::Application for Application {
+    fn update(&mut self, window_event: WindowEvent) {
+        match window_event {
+            WindowEvent::CursorMoved {
+                device_id,
+                position,
+            } => {
+                // Track positions.
+                //
+                // These positions aren't discrete / integral on macOS, but why?
+                let current = PointI::new(position.x.round() as _, position.y.round() as _);
+                self.positions.insert(device_id, current);
+
+                // Is there an ongoing movement on the left mouse button?
+                if let Some(pressed_state) = &self.left_mouse_button_pressed {
+                    let delta = current - pressed_state.origin;
+                    self.translation = pressed_state.translation_origin + delta;
+                }
+
+                if let Some(pressed_state) = &self.middle_mouse_button_pressed {
+                    let delta = current - pressed_state.origin;
+                    self.rotation = pressed_state.rotation_origin + delta;
+                }
+            }
+            WindowEvent::MouseWheel {
+                delta: MouseScrollDelta::PixelDelta(physical_position),
+                phase: TouchPhase::Moved,
+                ..
+            } => {
+                self.translation_z +=
+                    (physical_position.y * MOUSE_WHEEL_PIXEL_DELTA_TO_Z_PIXELS).round() as i32
+            }
+            WindowEvent::MouseWheel {
+                delta: MouseScrollDelta::LineDelta(_, y_delta),
+                phase: TouchPhase::Moved,
+                ..
+            } => self.translation_z += y_delta.round() as i32 * MOUSE_WHEEL_LINE_DELTA_TO_Z_PIXELS,
+            WindowEvent::MouseInput {
+                device_id,
+                state,
+                button: MouseButton::Left,
+            } if self.positions.contains_key(&device_id) => {
+                if state.is_pressed() {
+                    self.left_mouse_button_pressed = Some(LeftMouseButtonPressed {
+                        origin: self.positions[&device_id],
+                        translation_origin: self.translation,
+                    });
+                } else {
+                    self.left_mouse_button_pressed = None;
+                }
+            }
+            WindowEvent::MouseInput {
+                device_id,
+                state,
+                button: MouseButton::Middle,
+            } => {
+                if state.is_pressed() {
+                    self.middle_mouse_button_pressed = Some(MiddleMouseButtonPressed {
+                        origin: self.positions[&device_id],
+                        rotation_origin: self.rotation,
+                    });
+                } else {
+                    self.middle_mouse_button_pressed = None;
+                }
+            }
+
+            WindowEvent::MouseInput {
+                button: MouseButton::Right,
+                ..
+            } => {
+                self.rotation = PointI::default();
+            }
+            WindowEvent::ModifiersChanged(modifiers) => {
+                if self.modifiers != modifiers {
+                    // If there is an ongoing move and modifiers change, reset origins.
+                    // if let Some(ref mut mouse_pressed) = self.left_mouse_button_pressed {
+                    //     mouse_pressed.origin = self.positions[&mouse_pressed.device_id];
+                    //     mouse_pressed.translation_origin = self.translation;
+                    //     mouse_pressed.rotation_origin = self.rotation;
+                    // }
+
+                    self.modifiers = modifiers
+                }
+            }
+            WindowEvent::KeyboardInput {
+                event: KeyEvent {
+                    logical_key, state, ..
+                },
+                ..
+            } => {
+                if state == winit::event::ElementState::Pressed {
+                    match logical_key {
+                        Key::Named(NamedKey::ArrowLeft) => {
+                            self.camera.eye += Vector3::new(0.1, 0.0, 0.0)
+                        }
+                        Key::Named(NamedKey::ArrowRight) => {
+                            self.camera.eye -= Vector3::new(0.1, 0.0, 0.0)
+                        }
+                        Key::Named(NamedKey::ArrowUp) => {
+                            self.camera.eye += Vector3::new(0.0, 0.0, 0.1)
+                        }
+                        Key::Named(NamedKey::ArrowDown) => {
+                            self.camera.eye -= Vector3::new(0.0, 0.0, 0.1)
+                        }
+                        _ => {}
+                    }
+                } else {
+                    {}
+                }
+
+                println!("eye: {:?}", self.camera.eye)
+            }
+            _ => (),
+        }
+    }
+
+    fn render(&self, shell: &mut Shell) -> (Camera, Vec<Shape>) {
+        let mut shapes = Vec::new();
+
+        let page_x_center: f64 = -((self.page_size.width / 2) as f64);
+        let page_y_center: f64 = -((self.page_size.height / 2) as f64);
+        let center_transformation =
+            Matrix4::from_translation((page_x_center, page_y_center, 0.0).into());
+        let current_translation = Matrix4::from_translation(
+            (
+                self.translation.x as _,
+                self.translation.y as _,
+                self.translation_z as _,
+            )
+                .into(),
+        );
+        let angle_x = cgmath::Rad((self.rotation.x as f64 / 10.).to_radians());
+        let angle_y = cgmath::Rad((-self.rotation.y as f64 / 10.).to_radians());
+
+        let x_rotation = Matrix4::from_angle_y(angle_x);
+        let y_rotation = Matrix4::from_angle_x(angle_y);
+
+        let current_transformation =
+            current_translation * y_rotation * x_rotation * center_transformation;
+
+        for glyph_run in &self.glyph_runs {
+            // let center_x: i32 = (glyph_run.metrics.width / 2) as _;
+            // let center_y: i32 = ((glyph_run.metrics.size()).1 / 2) as _;
+
+            let local_offset = (glyph_run.0.x.floor(), glyph_run.0.y.floor());
+            let local_offset_matrix =
+                Matrix4::from_translation((local_offset.0, local_offset.1, 0.0).into());
+
+            let matrix = shell.pixel_matrix() * current_transformation * local_offset_matrix;
+
+            // TODO: Should we use `Rc` for GlyphRuns, too, so that that the application can keep them stored.
+            shapes.push(Shape::GlyphRun(matrix.into(), glyph_run.1.clone()));
+        }
+
+        (self.camera, shapes)
+    }
+}
